+++ plot/controls_test.go
@@ -1,17 +1,13 @@
 package plot_test
 
 import (
-	"math/rand"
-
 	"github.com/mlange-42/arche-model/model"
 	"github.com/mlange-42/arche-model/system"
 	"github.com/mlange-42/arche-pixel/plot"
 	"github.com/mlange-42/arche-pixel/window"
-	"github.com/mlange-42/arche/ecs"
 )
 
-func ExampleTimeSeries() {
-
+func ExampleControls() {
 	// Create a new model.
 	m := model.New()
 
@@ -19,12 +15,16 @@ func ExampleTimeSeries() {
 	m.TPS = 30
 	m.FPS = 0
 
-	// Create a time series plot.
-	// See below for the implementation of the RowObserver.
+	// Create a window with a Controls drawer.
 	m.AddUISystem((&window.Window{}).
-		With(&plot.TimeSeries{
-			Observer: &RowObserver{},
-		}))
+		With(&plot.Controls{Scale: 2}))
+
+	// Alternatively, add the Controls in addition to another drawer (e.g. Monitor).
+	m.AddUISystem((&window.Window{}).
+		With(
+			&plot.Monitor{},
+			&plot.Controls{},
+		))
 
 	// Add a termination system that ends the simulation.
 	m.AddSystem(&system.FixedTermination{
@@ -36,16 +36,4 @@ func ExampleTimeSeries() {
 	// Uncomment the next line. It is commented out as the CI has no display device to test the model run.
 
 	// pixelgl.Run(m.Run)
-}
-
-// RowObserver to generate random time series.
-type RowObserver struct{}
-
-func (o *RowObserver) Initialize(w *ecs.World) {}
-func (o *RowObserver) Update(w *ecs.World)     {}
-func (o *RowObserver) Header() []string {
-	return []string{"A", "B", "C"}
-}
-func (o *RowObserver) Values(w *ecs.World) []float64 {
-	return []float64{rand.Float64(), rand.Float64() + 1, rand.Float64() + 2}
 }